refactored

+++ schemas.py
@@ -1,10 +1,12 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class PhoneModel(BaseModel):
     id: str
     name: str
-    price: float
+    price: Decimal
     onMarket: bool
     description: str
     count: int
@@ -13,6 +15,6 @@ class PhoneModel(BaseModel):
 class FoodModel(BaseModel):
     id: str
     name: str
-    calories: float
+    calories: Decimal
     description: str
     count: int
